Split tests (#49)
