+++ Fambda.Tests/Core/BoolTypeTests.cs
