+++ Fambda.Tests/Core/Either/EitherLeftTests.cs
@@ -1,0 +1,55 @@
+using System;
+using Fambda.Tests.DataTypes;
+using FluentAssertions;
+using Microsoft.VisualStudio.TestTools.UnitTesting;
+
+using static Fambda.F;
+
+namespace Fambda.Tests
+{
+    [TestClass]
+    public class EitherLeftTests
+    {
+        [TestMethod]
+        public void EitherLeftCtorShouldSucceed()
+        {
+            // Arrange
+            var value = "left";
+
+            // Act
+            Action ctor = () => { new EitherLeft<string>(value); };
+
+            // Assert
+            ctor.Should().NotThrow();
+        }
+
+        [TestMethod]
+        public void EitherLeftShouldHaveExpectedValue()
+        {
+            // Arrange
+            var value = "left";
+            var eitherLeft = new EitherLeft<string>(value);
+
+            // Act
+            var result = eitherLeft.Value;
+
+            // Assert
+            result.Should().Be(value);
+        }
+
+        [TestMethod]
+        public void EitherLeftToStringShouldProvideExpectedRepresentation()
+        {
+            // Arrange
+            var value = "left";
+            var expectedResult = $"Left({value})";
+            var eitherLeft = new EitherLeft<string>(value);
+
+            // Act
+            var result = eitherLeft.ToString();
+
+            // Assert
+            result.Should().Be(expectedResult);
+        }
+    }
+}
