+++ Fambda.Tests/Core/Either/EitherRightTests.cs
@@ -1,0 +1,55 @@
+using System;
+using Fambda.Tests.DataTypes;
+using FluentAssertions;
+using Microsoft.VisualStudio.TestTools.UnitTesting;
+
+using static Fambda.F;
+
+namespace Fambda.Tests
+{
+    [TestClass]
+    public class EitherRightTests
+    {
+        [TestMethod]
+        public void EitherRightCtorShouldSucceed()
+        {
+            // Arrange
+            var value = "right";
+
+            // Act
+            Action ctor = () => { new EitherRight<string>(value); };
+
+            // Assert
+            ctor.Should().NotThrow();
+        }
+
+        [TestMethod]
+        public void EitherRightShouldHaveExpectedValue()
+        {
+            // Arrange
+            var value = "right";
+            var eitherRight = new EitherRight<string>(value);
+
+            // Act
+            var result = eitherRight.Value;
+
+            // Assert
+            result.Should().Be(value);
+        }
+
+        [TestMethod]
+        public void EitherRightToStringShouldProvideExpectedRepresentation()
+        {
+            // Arrange
+            var value = "right";
+            var expectedResult = $"Right({value})";
+            var eitherRight = new EitherRight<string>(value);
+
+            // Act
+            var result = eitherRight.ToString();
+
+            // Assert
+            result.Should().Be(expectedResult);
+        }
+    }
+}
