+++ Fambda.Tests/Core/Option/OptionTests.cs
@@ -1,0 +1,158 @@
+using System;
+using Fambda.Contracts;
+using FluentAssertions;
+using Microsoft.VisualStudio.TestTools.UnitTesting;
+
+using static Fambda.F;
+
+namespace Fambda.Tests
+{
+    [TestClass]
+    public class OptionTests
+    {
+        #region Option
+
+        [TestMethod]
+        public void OptionCtorThroughImplicitOperatorOverloadingShouldSucceedWithOptionSome()
+        {
+            // Arrange
+            var value = "value";
+            var optionSome = new OptionSome<string>(value);
+
+            // Act
+            Option<string> option = optionSome;
+            var result = option.ToString();
+
+            // Assert
+            result.Should().Be("Some(value)");
+        }
+
+        [TestMethod]
+        public void OptionCtorThroughImplicitOperatorOverloadingShouldSucceedWithOptionNone()
+        {
+            // Arrange
+            var optionNone = new OptionNone();
+
+            // Act
+            Option<string> option = optionNone;
+            var result = option.ToString();
+
+            // Assert
+            result.Should().Be("None");
+        }
+
+        [DataTestMethod]
+        [DataRow("value", "Some(value)")]
+        [DataRow(null, "None")]
+        public void OptionCtorThroughImplicitOperatorOverloadingShouldSucceedWithValue(string input, string expected)
+        {
+            // Arrange
+            var value = input;
+
+            // Act
+            Option<string> option = value;
+            var result = option.ToString();
+
+            // Assert
+            result.Should().Be(expected);
+        }
+
+        #endregion
+
+        #region OptionEquals
+
+        [TestMethod]
+        public void OptionEqualsOptionSomeShouldReturnTrue()
+        {
+            // Arrange
+            OptionSome<int> optionSome = new OptionSome<int>(1);
+            Option<int> option = Some(1);
+
+            // Act
+            var result = option.Equals(optionSome);
+
+            // Assert
+            result.Should().BeTrue();
+        }
+
+        [TestMethod]
+        public void OptionEqualsOptionSomeShouldReturnFalse()
+        {
+            // Arrange
+            OptionSome<int> optionSome = new OptionSome<int>(2);
+            Option<int> option = Some(1);
+
+            // Act
+            var result = option.Equals(optionSome);
+
+            // Assert
+            result.Should().BeFalse();
+        }
+
+        [TestMethod]
+        public void OptionEqualsOptionNoneShouldReturnTrue()
+        {
+            // Arrange
+            OptionNone optionNone = OptionNone.Default;
+            Option<int> option = None;
+
+            // Act
+            var result = option.Equals(optionNone);
+
+            // Assert
+            result.Should().BeTrue();
+        }
+
+        [TestMethod]
+        public void OptionEqualsOptionNoneShouldReturnFalse()
+        {
+            // Arrange
+            OptionNone optionNone = OptionNone.Default;
+            Option<int> option = Some(1);
+
+            // Act
+            var result = option.Equals(optionNone);
+
+            // Assert
+            result.Should().BeFalse();
+        }
+
+        #endregion
+
+        #region OptionMatch
+
+        [TestMethod]
+        public void OptionMatchShouldSucceedWhenSome()
+        {
+            // Arrange
+            Option<int> option = Some(1);
+
+            // Act
+            var result = option.Match(
+                                    Some: (i) => $"Result=Some({i})",
+                                    None: () => "Result=None"
+                                );
+
+            // Assert
+            result.Should().Be("Result=Some(1)");
+        }
+
+        [TestMethod]
+        public void OptionMatchShouldSucceedWhenNone()
+        {
+            // Arrange
+            Option<int> option = None;
+
+            // Act
+            var result = option.Match(
+                                    Some: (i) => $"Result=Some({i})",
+                                    None: () => "Result=None"
+                                );
+
+            // Assert
+            result.Should().Be("Result=None");
+        }
+
+        #endregion
+    }
+}
